+++ sdk/textanalytics/ai-text-analytics/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 6.0.0 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+- Extractive Summarization has been removed because it is still in preview.
+- Healthcare's FHIR has been removed because it is still in preview.
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 6.0.0-beta.1 (2022-05-19)
 
 This new major version beta introduces a full redesign of the Azure Text Analytics client library, supports the new Azure Cognitive Language Service API  (version "2022-04-01-preview" and newer) and drops support for the Text Analytics service API. Application code must be updated to use the new client, please see the [Migration Guide](https://github.com/Azure/azure-sdk-for-js/blob/main/sdk/textanalytics/ai-text-analytics/MIGRATION_v5_v6.md) for detailed instructions.